services/task-executor/requirements.txt
psycopg2-binary==2.8.2
redis==3.2.1